src/requirements.txt
requests>=2.30.0
prometheus_client>=0.16.0
python-dotenv>=1.0.0
opencode-ai>=0.1.0a36
pyyaml>=6.0
jsonschema>=4.0
discord.py>=2.3.0

# dev
pytest>=7.0.0
